Refactor to use Strategy Pattern for Parsing and Formatting
- I introduced a strategy pattern for parsing word data from HTML. This makes it easier to add new parsers for different parts of speech in the future.
- I introduced a strategy pattern for formatting word cards for display. This decouples the display logic from the handler and makes it easier to modify the appearance of word cards.
- I created new modules `app/services/parsing_strategies.py` and `app/handlers/card_formatters.py` to house the new strategies.
- I refactored `app/services/parser.py` and `app/handlers/common.py` to use the new strategies.
- I updated unit tests to reflect the changes in the codebase.
- I formatted code with `black`.
