+++ app/services/parser.py
@@ -1,330 +1,25 @@
 # -*- coding: utf-8 -*-
 
-import re
 import asyncio
-import unicodedata
 from typing import Tuple, Optional, Dict, Any, List
 from urllib.parse import quote, urljoin
 
 import httpx
-from bs4 import BeautifulSoup, Tag
-
+from bs4 import BeautifulSoup
 from pydantic import ValidationError
 
 from config import logger, PARSING_TIMEOUT
-from dal.unit_of_work import UnitOfWork
-from utils import normalize_hebrew, parse_translations
 from dal.models import CreateCachedWord, CreateTranslation, CreateVerbConjugation
+from dal.unit_of_work import UnitOfWork
+from services.parsing_strategies import (
+    get_parsing_strategy,
+    _get_part_of_speech_from_meta,
+)
+from utils import normalize_hebrew
 
 # --- УПРАВЛЕНИЕ КОНКУРЕНТНЫМ ПАРСИНГОМ ---
 PARSING_EVENTS: Dict[str, asyncio.Event] = {}
 PARSING_EVENTS_LOCK = asyncio.Lock()
-
-
-BINYAN_HTML_MAP = {
-    "пааль": "paal",
-    "пиэль": "piel",
-    "hифъиль": "hifil",
-    "нифъаль": "nifal",
-    "hитпаэль": "hitpael",
-    "hифъаль": "hufal",
-    "пуаль": "pual",
-}
-
-
-def _extract_form_value(cell: Tag) -> str:
-    """Извлекает иврит и транскрипцию из ячейки таблицы."""
-    menukad = cell.find(class_="menukad")
-    transcription = cell.find(class_="transcription")
-
-    hebrew_part = menukad.text.strip() if menukad else ""
-    trans_part = transcription.text.strip() if transcription else ""
-
-    if hebrew_part and trans_part:
-        return f"{hebrew_part} ({trans_part})"
-    return hebrew_part or trans_part
-
-
-def _extract_hebrew_from_cell(cell: Tag) -> Optional[str]:
-    """Извлекает только ивритскую форму из ячейки таблицы, без транскрипции."""
-    if not cell:
-        return None
-    menukad = cell.find(class_="menukad")
-    if not menukad:
-        return None
-    # Убираем альтернативные написания типа "~ ..."
-    hebrew_text = menukad.text.split("~")[0].strip()
-    # Нормализуем строку в NFD для консистентного сравнения
-    return unicodedata.normalize("NFD", hebrew_text)
-
-
-def _get_part_of_speech_from_meta(soup: BeautifulSoup) -> Optional[str]:
-    """
-    Извлекает часть речи из мета-тега description.
-    Это наиболее надежный метод.
-    """
-    meta_tag = soup.find("meta", attrs={"name": "description"})
-    if not meta_tag or not meta_tag.get("content"):
-        logger.warning("Мета-тег 'description' не найден или пуст.")
-        return None
-
-    content = meta_tag.get("content", "").lower()
-    if content.startswith("глагол"):
-        return "verb"
-    if content.startswith("существительное"):
-        return "noun"
-    if content.startswith("прилагательное"):
-        return "adjective"
-
-    logger.warning(f"Не удалось определить часть речи из мета-тега: {content[:50]}...")
-    return None
-
-
-def _parse_noun_forms(soup: BeautifulSoup) -> Dict[str, Any]:
-    """
-    Извлекает формы для существительного из таблицы.
-    Сохраняет род в стандартизированном английском формате.
-    """
-    forms = {}
-    main_header = soup.find("h2", class_="page-header")
-    if main_header:
-        p_tag = main_header.find_next_sibling("p")
-        if p_tag:
-            text = p_tag.text.lower()
-            if "мужской род" in text or "masculine" in text:
-                forms["gender"] = "masculine"
-            elif "женский род" in text or "feminine" in text:
-                forms["gender"] = "feminine"
-
-    table = soup.find("table", class_="conjugation-table")
-    if not table:
-        return forms
-
-    absolute_state_header = table.find("th", string="Абсолютное состояние")
-    if absolute_state_header:
-        row = absolute_state_header.parent
-        cells = row.find_all("td")
-        if len(cells) >= 2:
-            forms["singular_form"] = _extract_hebrew_from_cell(cells[0])
-            forms["plural_form"] = _extract_hebrew_from_cell(cells[1])
-
-    return forms
-
-
-def _parse_adjective_forms(soup: BeautifulSoup) -> Dict[str, Any]:
-    """
-    Извлекает формы для прилагательного из таблицы.
-    Теперь использует _extract_hebrew_from_cell для чистоты данных.
-    """
-    forms = {}
-    table = soup.find("table", class_="conjugation-table")
-    if not table:
-        return forms
-
-    data_row = table.find("tbody").find("tr")
-    if not data_row:
-        return forms
-
-    cells = data_row.find_all("td")
-    if len(cells) == 4:
-        forms["masculine_singular"] = _extract_hebrew_from_cell(cells[0])
-        forms["feminine_singular"] = _extract_hebrew_from_cell(cells[1])
-        forms["masculine_plural"] = _extract_hebrew_from_cell(cells[2])
-        forms["feminine_plural"] = _extract_hebrew_from_cell(cells[3])
-    return forms
-
-
-def parse_verb_page(soup: BeautifulSoup, main_header: Tag) -> Optional[Dict[str, Any]]:
-    """Парсер для страниц глаголов с детальным логированием."""
-    logger.info('{{"event": "parse_verb_page_start"}}')
-    try:
-        data = {"part_of_speech": "verb", "is_verb": True}
-
-        logger.debug('{{"event": "parsing_step", "step": "searching_infinitive"}}')
-        infinitive_div = soup.find("div", id="INF-L")
-
-        if not infinitive_div:
-            logger.error(
-                '{{"event": "parse_verb_error", "reason": "infinitive_block_not_found"}}'
-            )
-            return None
-
-        menukad_tag = infinitive_div.find("span", class_="menukad")
-        if not menukad_tag:
-            logger.error(
-                '{{"event": "parse_verb_error", "reason": "menukad_tag_not_found"}}'
-            )
-            return None
-
-        hebrew_text = menukad_tag.text.split("~")[0].strip()
-        data["hebrew"] = unicodedata.normalize("NFD", hebrew_text)
-        logger.debug(
-            f'{{"event": "parsing_step_success", "step": "infinitive_found", "hebrew": "{data["hebrew"]}"}}'
-        )
-
-        lead_div = soup.find("div", class_="lead")
-        if not lead_div:
-            logger.error(
-                f'{{"event": "parse_verb_error", "reason": "translation_div_not_found", "hebrew": "{data["hebrew"]}"}}'
-            )
-            return None
-        data["translations"] = parse_translations(lead_div.text.strip())
-        if not data["translations"]:
-            logger.warning(
-                f'{{"event": "parse_verb_warning", "reason": "empty_translations", "hebrew": "{data["hebrew"]}"}}'
-            )
-            return None
-
-        transcription_div = infinitive_div.find("div", class_="transcription")
-        data["transcription"] = (
-            transcription_div.text.strip() if transcription_div else ""
-        )
-
-        data["root"], data["binyan"] = None, None
-        for p in main_header.parent.find_all("p"):
-            text_lower = p.text.lower()
-            if "глагол" in text_lower:
-                binyan_text = (
-                    p.text.replace("Verb –", "").replace("Глагол –", "").strip()
-                )
-                binyan = binyan_text.split()[0] if binyan_text else None
-                binyan_enum_value = (
-                    BINYAN_HTML_MAP.get(binyan.lower()) if binyan else None
-                )
-                if binyan_enum_value:
-                    data["binyan"] = binyan_enum_value
-
-            if "root" in text_lower or "корень" in text_lower:
-                root_tag = p.find("span", class_="menukad")
-                if root_tag:
-                    data["root"] = root_tag.text.strip()
-
-        logger.debug(
-            f'{{"event": "parsing_step_success", "step": "root_and_binyan_parsed", "root": "{data["root"]}", "binyan": "{data["binyan"]}"}}'
-        )
-
-        conjugations = []
-        verb_forms = soup.find_all("div", id=re.compile(r"^(AP|PERF|IMPF|IMP|INF)-"))
-        for form in verb_forms:
-            form_id, menukad_tag, trans_tag = (
-                form.get("id"),
-                form.find("span", class_="menukad"),
-                form.find("div", class_="transcription"),
-            )
-            if all([form_id, menukad_tag, trans_tag]):
-                tense_prefix = form_id.split("-")[0].lower()
-                person = (
-                    form_id.split("-")[1] if len(form_id.split("-")) > 1 else "форма"
-                )
-                conjugations.append(
-                    {
-                        "tense": tense_prefix,
-                        "person": person,
-                        "hebrew_form": menukad_tag.text.strip().split("~")[0].strip(),
-                        "transcription": trans_tag.text.strip(),
-                    }
-                )
-        data["conjugations"] = conjugations
-
-        logger.info(
-            f'{{"event": "parse_verb_page_success", "found_conjugations": {len(conjugations)}}}'
-        )
-        return data
-    except Exception as e:
-        logger.error(
-            f'{{"event": "parse_verb_page_exception", "error": "{e}"}}',
-            exc_info=True,
-        )
-        return None
-
-
-def parse_noun_or_adjective_page(
-    soup: BeautifulSoup, main_header: Tag
-) -> Optional[Dict[str, Any]]:
-    """
-    Парсер для страниц существительных и прилагательных с детальным логированием.
-    Ищет каноническую форму непосредственно в таблице склонения.
-    """
-    part_of_speech = _get_part_of_speech_from_meta(soup)
-    if not part_of_speech or part_of_speech == "verb":
-        logger.error(
-            f'{{"event": "wrong_parser_error", "part_of_speech": "{part_of_speech}"}}'
-        )
-        return None
-
-    logger.info(
-        f'{{"event": "parse_noun_adj_start", "part_of_speech": "{part_of_speech}"}}'
-    )
-    try:
-        data = {
-            "root": None,
-            "binyan": None,
-            "conjugations": [],
-            "part_of_speech": part_of_speech,
-            "is_verb": False,
-        }
-
-        logger.debug('{{"event": "parsing_step", "step": "searching_canonical_form"}}')
-        canonical_hebrew = None
-        table = soup.find("table", class_="conjugation-table")
-
-        if table:
-            if part_of_speech == "noun":
-                header_cell = table.find("th", string="Абсолютное состояние")
-                if header_cell:
-                    data_cell = header_cell.parent.find("td")
-                    if data_cell:
-                        canonical_hebrew = _extract_hebrew_from_cell(data_cell)
-            elif part_of_speech == "adjective":
-                data_cell = table.find("td")
-                if data_cell:
-                    canonical_hebrew = _extract_hebrew_from_cell(data_cell)
-
-        if not canonical_hebrew:
-            logger.error(
-                '{{"event": "parse_noun_adj_error", "reason": "canonical_form_not_found"}}'
-            )
-            return None
-
-        data["hebrew"] = canonical_hebrew
-        logger.debug(
-            f'{{"event": "parsing_step_success", "step": "canonical_form_found", "hebrew": "{data["hebrew"]}"}}'
-        )
-
-        lead_div = soup.find("div", class_="lead")
-        if not lead_div:
-            logger.error(
-                f'{{"event": "parse_noun_adj_error", "reason": "translation_div_not_found", "hebrew": "{data["hebrew"]}"}}'
-            )
-            return None
-
-        data["translations"] = parse_translations(lead_div.text.strip())
-        if not data["translations"]:
-            logger.warning(
-                f'{{"event": "parse_noun_adj_warning", "reason": "empty_translations", "hebrew": "{data["hebrew"]}"}}'
-            )
-            return None
-
-        first_form = soup.find("div", class_="transcription")
-        data["transcription"] = first_form.text.strip() if first_form else ""
-
-        if data["part_of_speech"] == "noun":
-            forms = _parse_noun_forms(soup)
-            data.update(forms)
-        elif data["part_of_speech"] == "adjective":
-            forms = _parse_adjective_forms(soup)
-            data.update(forms)
-
-        logger.info(
-            f'{{"event": "parse_noun_adj_success", "hebrew": "{data["hebrew"]}"}}'
-        )
-        return data
-    except Exception as e:
-        logger.error(
-            f'{{"event": "parse_noun_adj_exception", "error": "{e}"}}',
-            exc_info=True,
-        )
-        return None
 
 
 async def _parse_disambiguation_page(
@@ -374,8 +69,7 @@ async def _parse_disambiguation_page(
 
 def _parse_single_word_page(soup: BeautifulSoup) -> Optional[Dict]:
     """
-    Определяет тип страницы (глагол, существительное и т.д.) и вызывает
-    соответствующий парсер, добавляя логирование на каждом шаге.
+    Определяет тип страницы, выбирает стратегию парсинга и обрабатывает результат.
     """
     logger.info('{{"event": "determine_page_type"}}')
 
@@ -386,27 +80,24 @@ def _parse_single_word_page(soup: BeautifulSoup) -> Optional[Dict]:
         )
         return None
 
-    parsed_data = None
     part_of_speech = _get_part_of_speech_from_meta(soup)
-
-    if (
-        part_of_speech == "verb"
-        or "спряжение" in main_header.text.lower()
+    # Fallback for verbs if meta tag is missing
+    if not part_of_speech and (
+        "спряжение" in main_header.text.lower()
         or "conjugation" in main_header.text.lower()
     ):
-        logger.info('{{"event": "page_type_determined", "type": "verb"}}')
-        parsed_data = parse_verb_page(soup, main_header)
-    elif (
-        part_of_speech in ["noun", "adjective"]
-        or "формы слова" in main_header.text.lower()
-    ):
-        logger.info(f'{{"event": "page_type_determined", "type": "{part_of_speech}"}}')
-        parsed_data = parse_noun_or_adjective_page(soup, main_header)
-    else:
+        part_of_speech = "verb"
+
+    strategy = get_parsing_strategy(part_of_speech)
+
+    if not strategy:
         logger.error(
             f'{{"event": "page_type_error", "reason": "unknown_page_type", "header": "{main_header.text.strip()}"}}'
         )
         return None
+
+    logger.info(f'{{"event": "page_type_determined", "type": "{part_of_speech}"}}')
+    parsed_data = strategy.parse(soup, main_header)
 
     if not parsed_data:
         logger.error(
@@ -418,30 +109,39 @@ def _parse_single_word_page(soup: BeautifulSoup) -> Optional[Dict]:
         f'{{"event": "parsing_successful", "hebrew": "{parsed_data.get("hebrew", "N/A")}"}}'
     )
 
-    # Нормализация и подготовка данных к сохранению
-    parsed_data["normalized_hebrew"] = normalize_hebrew(parsed_data["hebrew"])
-    if parsed_data.get("conjugations"):
-        for conj in parsed_data["conjugations"]:
-            conj["normalized_hebrew_form"] = normalize_hebrew(conj["hebrew_form"])
+    return _create_word_model_from_parsed_data(parsed_data)
 
-    parsed_translations = [
-        CreateTranslation(**t) for t in parsed_data.get("translations", [])
-    ]
-    parsed_conjugations = [
-        CreateVerbConjugation(**c) for c in parsed_data.get("conjugations", [])
-    ]
 
+def _create_word_model_from_parsed_data(
+    parsed_data: Dict[str, Any],
+) -> Optional[CreateCachedWord]:
+    """Нормализует и валидирует сырые данные парсинга, создавая модель слова."""
     try:
-        # Полная и явная инициализация модели
-        word_model = CreateCachedWord(
-            hebrew=parsed_data["hebrew"],
-            normalized_hebrew=parsed_data["normalized_hebrew"],
+        hebrew = parsed_data.get("hebrew")
+        if not hebrew:
+            logger.error("Critical error: parsed_data is missing 'hebrew' key.")
+            return None
+
+        normalized_hebrew = normalize_hebrew(hebrew)
+        parsed_data["normalized_hebrew"] = normalized_hebrew
+
+        if parsed_data.get("conjugations"):
+            for conj in parsed_data["conjugations"]:
+                conj["normalized_hebrew_form"] = normalize_hebrew(conj["hebrew_form"])
+
+        return CreateCachedWord(
+            hebrew=hebrew,
+            normalized_hebrew=normalized_hebrew,
             transcription=parsed_data.get("transcription"),
             part_of_speech=parsed_data.get("part_of_speech"),
             root=parsed_data.get("root"),
             binyan=parsed_data.get("binyan"),
-            translations=parsed_translations,
-            conjugations=parsed_conjugations,
+            translations=[
+                CreateTranslation(**t) for t in parsed_data.get("translations", [])
+            ],
+            conjugations=[
+                CreateVerbConjugation(**c) for c in parsed_data.get("conjugations", [])
+            ],
             gender=parsed_data.get("gender"),
             singular_form=parsed_data.get("singular_form"),
             plural_form=parsed_data.get("plural_form"),
@@ -450,8 +150,6 @@ def _parse_single_word_page(soup: BeautifulSoup) -> Optional[Dict]:
             masculine_plural=parsed_data.get("masculine_plural"),
             feminine_plural=parsed_data.get("feminine_plural"),
         )
-        return word_model
-
     except ValidationError as e:
         logger.error(
             f"Ошибка валидации данных после парсинга для слова '{parsed_data.get('hebrew')}': {e}"
